Rewrite of querySelector functions into S3 methods.
Enables support of xml2 documents in addition to the existing support
for XML documents.

Additionally, XML and xml2 are now optional and will only be loaded if
present on the existing system. This will avoid the need for any large
dependencies to be installed if the user only wants to perform
translation (e.g. rvest).