Attempt to fix 'method missing' errors on ref classes.
This appears to be due to late binding of methods as methods are
bound when the '$' operator is used. See .dollarForEnvRefClass and
envRefInferField in refClass.R (methods package).

The attempted fix is to manually bind the method to the '.self' object
and call that method instead of calling by name on the object.